requirements.txt
django
psycopg2 >= 2.6.2
DateTime==4.0.1
PyBrain==0.3
dota2py==0.1.3
threadpool==1.3.2
requests
mock==2.0.0
redis==2.10.3
celery==3.1.18
gunicorn
djangorestframework
